Fix image URLs
When you run this script on your local machine, everything orks perfect. 
But from my experience, when uploading the files to the server, for some reason the server is Capital Sensitive, meaning that if the actual path to the image is 
img/Neymar.jpg

Trying to access it through "img/neymar.jpg" will only work on your local machine.
Therefore, I removed the .toLowerCase() method from the image URL. :)

Please note that I havn't tested it, I'm only guessing.